stylecheck web app/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
mistralai
anthropic
openai
google-generativeai
python-dotenv
